requirements.txt
absl-py==2.1.0
anyio==4.6.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.5.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
autograd==1.7.0
autoray==0.6.12
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
cachetools==5.5.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
chex==0.1.87
click==8.0.3
colorama==0.4.4
comm==0.2.2
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
dm-tree==0.1.8
docstring_parser==0.16
etils==1.10.0
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.14.1
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
flax==0.10.0
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.2.0
future==0.18.2
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.2
humanize==4.11.0
idna==3.3
imageio==2.36.0
immutabledict==4.2.0
importlib-metadata==4.6.4
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
ipython_genutils==0.2.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.6.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
medmnist==2.0.1
mistune==3.0.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optax==0.2.3
optree==0.13.0
orbax-checkpoint==0.7.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
PennyLane==0.38.0
PennyLane_Lightning==0.38.0
pexpect==4.8.0
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.0.0
python-json-logger==2.0.7
pytz==2022.1
pyxdg==0.27
pyzmq==26.2.0
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
rustworkx==0.15.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
Send2Trash==1.8.3
simple-parsing==0.1.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorcircuit==0.12.0
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensornetwork-ng==0.5.0
tensorstore==0.1.66
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.1.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.5
virtualenv==20.26.6
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
xdg==5
zipp==1.0.0


